chore: Open modal with a button
