+++ test/components/viral/modal/content_component_test.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require "test_helper"
+
+class Viral::Modal::ContentComponentTest < ViewComponent::TestCase
+  def test_component_renders_something_useful
+    # assert_equal(
+    #   %(<span>Hello, components!</span>),
+    #   render_inline(Viral::Modal::ContentComponent.new(message: "Hello, components!")).css("span").to_html
+    # )
+  end
+end
